Various medium priority bonus changes/additions
Added a url parameter "loc". The map will scan for this parameter and only display companies with an address containing the value.
Added contributors Chris Cera, Brad Oyler, and Andrew Scott.
Fixed a bug in the map that caused broken links to be displayed.
Added a "Jobs Page" url to companies that gets displayed on the map.
Added category filtering to the map widget.